chat_template/requirements.txt
python-fasthtml
uvicorn>=0.29
google-cloud-aiplatform